Move CPC Char into uniquecount, add CPC speed char.
