--- src/main/java/com/yahoo/sketches/characterization/cpc/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright 2018, Yahoo! Inc. Licensed under the terms of the
- * Apache License 2.0. See LICENSE file at the project root for terms.
- */
-
-/**
- * @author Lee Rhodes
- */
-
-package com.yahoo.sketches.characterization.cpc;
