+++ src/main/java/com/yahoo/sketches/characterization/uniquecount/CpcMergingValidationProfile.java
@@ -3,7 +3,7 @@
  * Apache License 2.0. See LICENSE file at the project root for terms.
  */
 
-package com.yahoo.sketches.characterization.cpc;
+package com.yahoo.sketches.characterization.uniquecount;
 
 import java.io.PrintStream;
 import java.io.PrintWriter;
@@ -11,19 +11,17 @@ import java.io.PrintWriter;
 import com.yahoo.sketches.characterization.Job;
 import com.yahoo.sketches.characterization.JobProfile;
 import com.yahoo.sketches.characterization.Properties;
-import com.yahoo.sketches.cpc.CompressionCharacterization;
+import com.yahoo.sketches.cpc.MergingValidation;
 
 /**
  * @author Lee Rhodes
  */
-public class CpcCompressionCharacterizationProfile implements JobProfile {
+public class CpcMergingValidationProfile implements JobProfile {
   Job job;
   Properties prop;
 
   int lgMinK;
   int lgMaxK; //inclusive
-  int lgMinT;
-  int lgMaxT;
   int lgMulK;
   int uPPO;
   int incLgK;
@@ -37,8 +35,6 @@ public class CpcCompressionCharacterizationProfile implements JobProfile {
 
     lgMinK = Integer.parseInt(prop.mustGet("lgMinK"));
     lgMaxK = Integer.parseInt(prop.mustGet("lgMaxK"));
-    lgMinT = Integer.parseInt(prop.mustGet("lgMinT"));
-    lgMaxT = Integer.parseInt(prop.mustGet("lgMaxT"));
     lgMulK = Integer.parseInt(prop.mustGet("lgMulK"));
     uPPO = Integer.parseInt(prop.mustGet("uPPO"));
     incLgK = Integer.parseInt(prop.mustGet("incLgK"));
@@ -46,9 +42,9 @@ public class CpcCompressionCharacterizationProfile implements JobProfile {
     ps = System.out;
     pw = job.getPrintWriter();
 
-    final CompressionCharacterization cc = new CompressionCharacterization(
-        lgMinK, lgMaxK, lgMinT, lgMaxT, lgMulK, uPPO, incLgK, ps, pw);
-    cc.start();
+    final MergingValidation mv = new MergingValidation(
+        lgMinK, lgMaxK, lgMulK, uPPO, incLgK, ps, pw);
+    mv.start();
   }
 
   @Override
