+++ src/main/java/com/yahoo/sketches/characterization/uniquecount/CpcQuickMergingValidationProfile.java
@@ -3,7 +3,7 @@
  * Apache License 2.0. See LICENSE file at the project root for terms.
  */
 
-package com.yahoo.sketches.characterization.cpc;
+package com.yahoo.sketches.characterization.uniquecount;
 
 import java.io.PrintStream;
 import java.io.PrintWriter;
@@ -11,19 +11,18 @@ import java.io.PrintWriter;
 import com.yahoo.sketches.characterization.Job;
 import com.yahoo.sketches.characterization.JobProfile;
 import com.yahoo.sketches.characterization.Properties;
-import com.yahoo.sketches.cpc.MergingValidation;
+import com.yahoo.sketches.cpc.QuickMergingValidation;
 
 /**
  * @author Lee Rhodes
  */
-public class CpcMergingValidationProfile implements JobProfile {
+public class CpcQuickMergingValidationProfile implements JobProfile {
+
   Job job;
   Properties prop;
 
   int lgMinK;
   int lgMaxK; //inclusive
-  int lgMulK;
-  int uPPO;
   int incLgK;
   PrintStream ps;
   PrintWriter pw;
@@ -35,15 +34,13 @@ public class CpcMergingValidationProfile implements JobProfile {
 
     lgMinK = Integer.parseInt(prop.mustGet("lgMinK"));
     lgMaxK = Integer.parseInt(prop.mustGet("lgMaxK"));
-    lgMulK = Integer.parseInt(prop.mustGet("lgMulK"));
-    uPPO = Integer.parseInt(prop.mustGet("uPPO"));
     incLgK = Integer.parseInt(prop.mustGet("incLgK"));
 
     ps = System.out;
     pw = job.getPrintWriter();
 
-    final MergingValidation mv = new MergingValidation(
-        lgMinK, lgMaxK, lgMulK, uPPO, incLgK, ps, pw);
+    final QuickMergingValidation mv = new QuickMergingValidation(
+        lgMinK, lgMaxK, incLgK, ps, pw);
     mv.start();
   }
 
@@ -55,5 +52,4 @@ public class CpcMergingValidationProfile implements JobProfile {
 
   @Override
   public void println(final String s) {}
-
 }
