+++ src/main/resources/uniquecount/CpcCompressionCharacterizationJob.conf
