+++ src/main/resources/uniquecount/CpcUpdateSpeedJob.conf
@@ -1,0 +1,24 @@
+# Job
+
+#Uniques Profile
+Trials_lgMinU=0  #The starting # of uniques that is printed at the end.
+Trials_lgMaxU=20 #How high the # uniques go
+Trials_UPPO=16   #The horizontal x-resolution of trials points
+
+# Trials Profile
+Trials_lgMaxT=23 #Max trials at start (low counts)
+Trials_lgMinT=4  #Min trials at tail (high counts) 
+
+#Trails Speed related
+Trials_lgMinBpU=4   #start the downward slope of trials at this LgU
+Trials_lgMaxBpU=20  #stop the downward slope of trials at this LgU
+
+# Date-Time Profile
+TimeZone=PST
+TimeZoneOffset=-28800000 # offset in millisec
+FileNameDateFormat=yyyyMMdd'_'HHmmssz
+ReadableDateFormat=yyyy/MM/dd HH:mm:ss z
+
+#Job Profile
+JobProfile=com.yahoo.sketches.characterization.uniquecount.CpcUpdateSpeedProfile
+LgK=12
